+++ src/client/protocols/protocolgame.cpp
@@ -20,24 +20,43 @@
  * THE SOFTWARE.
  */
 
-#ifndef PROTOCOLCODES_H
-#define PROTOCOLCODES_H
+#include "protocolgame.h"
+#include "../game.h"
+#include "../item.h"
+#include "../localplayer.h"
+#include "../player.h"
 
-#include "global.h"
+void ProtocolGame::login(const std::string& accountName, const std::string& accountPassword, const std::string& host, uint16 port, const std::string& characterName, const std::string& authenticatorToken, const std::string& sessionKey)
+{
+    m_accountName = accountName;
+    m_accountPassword = accountPassword;
+    m_authenticatorToken = authenticatorToken;
+    m_sessionKey = sessionKey;
+    m_characterName = characterName;
 
-namespace Proto {
-    enum CreaturesIdRange {
-        PlayerStartId = 0x10000000,
-        PlayerEndId = 0x40000000,
-        MonsterStartId = 0x40000000,
-        MonsterEndId = 0x80000000,
-        NpcStartId = 0x80000000,
-        NpcEndId = 0xffffffff
-    };
-
-    void buildMessageModesMap(int version);
-    Otc::MessageMode translateMessageModeFromServer(uint8 mode);
-    uint8 translateMessageModeToServer(Otc::MessageMode mode);
+    connect(host, port);
 }
 
-#endif
+void ProtocolGame::onConnect()
+{
+    Protocol::onConnect();
+
+    m_localPlayer = g_game.getLocalPlayer();
+
+    if(!g_game.getFeature(Otc::GameChallengeOnLogin))
+        sendLoginPacket(0, 0);
+
+    recv();
+}
+
+void ProtocolGame::onRecv(const InputMessagePtr& inputMessage)
+{
+    parseMessage(inputMessage);
+    recv();
+}
+
+void ProtocolGame::onError(const boost::system::error_code& error)
+{
+    g_game.processConnectionError(error);
+    disconnect();
+}
