+++ src/client/protocols/protocolgamesend.cpp
@@ -23,8 +23,8 @@
 #include <framework/core/application.h>
 #include <framework/platform/platform.h>
 #include <framework/util/crypt.h>
-#include "client.h"
-#include "game.h"
+#include "../client.h"
+#include "../game.h"
 #include "protocolgame.h"
 
 void ProtocolGame::send(const OutputMessagePtr& outputMessage, bool skipXtea)
